Cache preferences and services in status dialog.
Currently, the status dialog fetches the preferences and system
services that it uses to display connectivity information (e.g.,
IP and MAC addresses) only when the dialog is first opened. This
is fine because the connectivity information is only filled in
once and never updated.

In order to have this information update dynamically (like the
rest of the dialog), fetch and cache these objects in onCreate.
The next change will actually make this information update on
network changes.

[Cherry-pick of 6eb6a908217e080aea43406028be491d269bdbb0]

Bug: 10232006
Change-Id: Ib4072674543e517cf6935e3e03f35090e728090a